+++ pylatex/math.py
@@ -39,7 +39,7 @@ class Math(Container):
 
         if self.inline:
             return '$' + self.dumps_content() + '$'
-        return '$$\n' + self.dumps_content() + '\n$$'
+        return '$$%\n' + self.dumps_content() + '%\n$$'
 
 
 class VectorName(Command):
@@ -118,7 +118,7 @@ class Matrix(Environment):
             string += str(value)
 
             if x == shape[1] - 1 and y != shape[0] - 1:
-                string += r'\\' + '\n'
+                string += r'\\' + '%\n'
 
         super().dumps_content()